fix: rename step to 'jx step verify behavior'
